+++ CHANGELOG.md
@@ -40,7 +40,7 @@
 
 * New `expressPreSession` server-side hook.
 * Pad server-side hook changes:
-  * `padCopy`: New `srcPad` context property.
+  * `padCopy`: New `srcPad` and `dstPad` context properties.
   * `padDefaultContent`: New hook.
 * The `db` property on Pad objects is now public.
 * New `getAuthorId` server-side hook.
@@ -68,8 +68,10 @@
   `handleMessage` server-side hooks is deprecated; use the `socket` context
   property instead.
 * Pad server-side hook changes:
-  * `padCopy`: The `originalPad` context property is deprecated; use `srcPad`
-    instead.
+  * `padCopy`:
+    * The `originalPad` context property is deprecated; use `srcPad` instead.
+    * The `destinationID` context property is deprecated; use `dstPad.id`
+      instead.
   * `padCreate`: The `author` context property is deprecated; use the new
     `authorId` context property instead.
   * `padUpdate`: The `author` context property is deprecated; use the new